refactor(img): rename Img → Image
